v2/pendle_fixed_yield/python/pyproject.toml
[project]
name = "pendle_fixed_yield_python_example"
version = "1.0.0"
description = "Example: Deposit into Pendle Fixed Yield Position using Compass API"
dependencies = [
    "compass-api-sdk",
    "python-dotenv",
    "web3",
    "eth-account",
]
